requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=4.0.0=pypi_0
asgiref=3.7.2=pypi_0
astor=0.8.1=pypi_0
babel=2.13.0=pypi_0
bce-python-sdk=0.8.92=pypi_0
blas=1.0=mkl
blinker=1.6.3=pypi_0
brotlipy=0.7.0=py311h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2023.08.22=haa95532_0
certifi=2023.7.22=py311haa95532_0
cffi=1.15.1=py311h2bbff1b_3
charset-normalizer=3.3.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.1.1=pypi_0
cryptography=41.0.3=py311h89fc84f_0
cuda-cccl=12.2.140=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
django=4.2.6=pypi_0
expat=2.5.0=hd77b12b_0
filelock=3.12.4=pypi_0
flask=3.0.0=pypi_0
flask-babel=4.0.0=pypi_0
fonttools=4.43.1=pypi_0
freetype=2.12.1=ha860e81_0
future=0.18.3=pypi_0
giflib=5.2.1=h8cc25b3_3
h11=0.14.0=pypi_0
httpcore=0.18.0=pypi_0
httpx=0.25.0=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.4=py311haa95532_0
intel-openmp=2023.1.0=h59b6b97_46319
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=py311haa95532_0
joblib=1.3.2=pypi_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.5=pypi_0
lerc=3.0=hd77b12b_0
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.3.141=0
libcurand-dev=10.3.3.141=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.0=pypi_0
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.8=py311h2bbff1b_0
mkl_random=1.2.4=py311h59b6b97_0
mpmath=1.3.0=py311haa95532_0
networkx=3.1=py311haa95532_0
numpy=1.26.1=pypi_0
numpy-base=1.24.3=py311hd01c5d8_1
opencv-python=4.5.5.64=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.11=h2bbff1b_2
opt-einsum=3.3.0=pypi_0
packaging=23.2=pypi_0
paddle-bfloat=0.1.7=pypi_0
paddlepaddle-gpu=2.5.1.post117=pypi_0
pandas=2.1.1=pypi_0
pillow=10.1.0=pypi_0
pip=23.2.1=py311haa95532_0
prettytable=3.9.0=pypi_0
protobuf=3.20.2=pypi_0
psutil=5.9.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.19.0=pypi_0
pyopenssl=23.2.0=py311haa95532_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py311haa95532_0
python=3.11.5=he1021f5_0
python-dateutil=2.8.2=pypi_0
pytorch=2.0.1=py3.11_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
rarfile=4.1=pypi_0
requests=2.31.0=py311haa95532_0
scikit-learn=1.3.0=py311hf62ec03_0
scipy=1.11.3=pypi_0
setuptools=68.0.0=py311haa95532_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
sqlparse=0.4.4=pypi_0
sympy=1.11.1=py311haa95532_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h2bbff1b_0
torchaudio=2.0.2=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.66.1=pypi_0
typing_extensions=4.7.1=py311haa95532_0
tzdata=2023.3=pypi_0
urllib3=2.0.7=pypi_0
vc=14.2=h21ff451_1
visualdl=2.5.3=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.8=pypi_0
werkzeug=3.0.0=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py311haa95532_0
xz=5.4.2=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
